updates 3/4/2022

+++ DBGeneration/Models/CorporateRejectedReasonHistory.cs
@@ -5,8 +5,8 @@ namespace DBGeneration.Models
 {
     public partial class CorporateRejectedReasonHistory
     {
-        public int ID { get; set; }
-        public Nullable<int> CorporateID { get; set; }
+        public long ID { get; set; }
+        public Nullable<long> CorporateID { get; set; }
         public string RejectedReason { get; set; }
         public string Action { get; set; }
         public Nullable<int> ActionBy { get; set; }
