+++ DBGeneration/Models/CustomerRejectedReasonHistory.cs
@@ -5,8 +5,8 @@ namespace DBGeneration.Models
 {
     public partial class CustomerRejectedReasonHistory
     {
-        public int ID { get; set; }
-        public Nullable<int> CustomerID { get; set; }
+        public long ID { get; set; }
+        public Nullable<long> CustomerID { get; set; }
         public string RejectedReason { get; set; }
         public string Action { get; set; }
         public Nullable<int> ActionBy { get; set; }
