+++ Inno.Core/Data/CorporateStatusHistory.cs
@@ -5,8 +5,8 @@ namespace Inno.Core.Data
 {
     public partial class CorporateStatusHistory:BaseEntity
     {
-        public int ID { get; set; }
-        public Nullable<int> CorporateID { get; set; }
+        public long ID { get; set; }
+        public Nullable<long> CorporateID { get; set; }
         public Nullable<int> CustomerStatusID { get; set; }
         public string Action { get; set; }
         public Nullable<int> ActionBy { get; set; }
