+++ Inno.Core/Data/Customer.cs
@@ -5,7 +5,7 @@ namespace Inno.Core.Data
 {
     public partial class Customer:BaseEntity
     {
-        public int ID { get; set; }
+        public long ID { get; set; }
         public string RequestNumber { get; set; }
         public string Name { get; set; }
         public string Address { get; set; }
