+++ Inno.Core/Data/CustomerRejectedReasonHistory.cs
@@ -5,8 +5,8 @@ namespace Inno.Core.Data
 {
     public partial class CustomerRejectedReasonHistory : BaseEntity
     {
-        public int ID { get; set; }
-        public Nullable<int> CustomerID { get; set; }
+        public long ID { get; set; }
+        public Nullable<long> CustomerID { get; set; }
         public string RejectedReason { get; set; }
         public string Action { get; set; }
         public Nullable<int> ActionBy { get; set; }
